+++ server/start.go
@@ -149,13 +149,16 @@ which accepts a path for the resulting pprof file.
 			serverCtx.Logger.Info("Unlocking keyring")
 
 			// fire unlock precess for keyring
-			keyringBackend, _ := cmd.Flags().GetString(flags.FlagKeyringBackend)
-			if keyringBackend == keyring.BackendFile {
+			krBackend := clientCtx.Keyring.Backend()
+			if krBackend == keyring.BackendFile {
 				_, err = clientCtx.Keyring.List()
 				if err != nil {
 					return err
 				}
 			}
+
+			// set keyring backend type to the server context
+			serverCtx.Viper.Set(flags.FlagKeyringBackend, krBackend)
 
 			serverCtx.Logger.Info("starting ABCI with CometBFT")